Extract plugin-api-blockproduction module
Move the ten transaction selection and mining control contracts into
ethereum/blockcreation/plugin-api, published as
besu-plugin-api-blockproduction and re-exported by the umbrella, so the
surface reachable through besu-plugin-api is unchanged. Packages are
untouched. MiningServiceImpl and TransactionSelectionServiceImpl move to
ethereum.blockcreation.pluginadapter. Javadoc-only fixes ride the move.

Fixes #11151
