+++ app/src/main/java/org/hyperledger/besu/services/BesuPluginServiceRegistrar.java
@@ -19,6 +19,7 @@ import org.hyperledger.besu.controller.BesuController;
 import org.hyperledger.besu.ethereum.api.pluginadapter.HealthCheckServiceImpl;
 import org.hyperledger.besu.ethereum.api.pluginadapter.TraceServiceImpl;
 import org.hyperledger.besu.ethereum.api.query.BlockchainQueries;
+import org.hyperledger.besu.ethereum.blockcreation.pluginadapter.MiningServiceImpl;
 import org.hyperledger.besu.ethereum.chain.pluginadapter.RlpConverterServiceImpl;
 import org.hyperledger.besu.ethereum.core.MiningConfiguration;
 import org.hyperledger.besu.ethereum.eth.transactions.pluginadapter.TransactionPoolServiceImpl;
@@ -213,8 +214,7 @@ public final class BesuPluginServiceRegistrar {
                 miningConfiguration),
             besuController.getProtocolSchedule()));
 
-    pluginContext.addService(
-        MiningService.class, new MiningServiceImpl(besuController.getMiningCoordinator()));
+    registerMiningService(pluginContext, besuController);
 
     pluginContext.addService(
         BlockSimulationService.class,
@@ -226,5 +226,12 @@ public final class BesuPluginServiceRegistrar {
             besuController.getProtocolContext().getBlockchain()));
 
     besuController.getAdditionalPluginServices().appendPluginServices(pluginContext);
+  }
+
+  @SuppressWarnings("removal") // MiningService is deprecated for removal; drop this with it
+  private static void registerMiningService(
+      final BesuPluginContextImpl pluginContext, final BesuController besuController) {
+    pluginContext.addService(
+        MiningService.class, new MiningServiceImpl(besuController.getMiningCoordinator()));
   }
 }
