+++ ethereum/blockcreation/plugin-api/build.gradle
@@ -1,0 +1,57 @@
+/*
+ * Copyright contributors to Besu.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+apply plugin: 'java-library'
+
+// override the path-derived name (besu-ethereum-blockcreation-plugin-api) with the plugin API family naming
+def blockproductionApiArtifactId = 'besu-plugin-api-blockproduction'
+
+jar {
+  archiveBaseName = blockproductionApiArtifactId
+  manifest {
+    attributes(
+      'Specification-Title': archiveBaseName,
+      'Specification-Version': project.version,
+      'Implementation-Title': archiveBaseName,
+      'Implementation-Version': calculateVersion(),
+      'Commit-Hash': getGitCommitDetails(40).hash,
+      'Automatic-Module-Name': 'org.hyperledger.besu.plugin.api.blockproduction'
+      )
+  }
+}
+
+dependencies {
+  api project(':plugin-api:core')
+  // PendingTransaction and Wei appear in this module's own signatures, so the
+  // dependency is declared here even though plugin-api:core also re-exports it
+  api project(':datatypes')
+  // PluginTransactionSelector.getOperationTracer returns BlockAwareOperationTracer
+  api project(':ethereum:core:plugin-api-execution')
+  // TransactionEvaluationContext.getEvaluationTimer returns a guava Stopwatch
+  api 'com.google.guava:guava'
+}
+
+publishing {
+  publications {
+    mavenJava(MavenPublication) {
+      groupId = 'org.hyperledger.besu'
+      artifactId = blockproductionApiArtifactId
+      pom {
+        name = 'Besu Plugin API Block Production'
+        description = 'Transaction selection contracts of the Besu plugin API, letting plugins filter and propose the transactions of the blocks the node produces'
+      }
+    }
+  }
+}
