+++ ethereum/blockcreation/plugin-api/src/main/java/org/hyperledger/besu/plugin/data/TransactionSelectionResult.java
