+++ ethereum/blockcreation/src/main/java/org/hyperledger/besu/ethereum/blockcreation/pluginadapter/TransactionSelectionServiceImpl.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.services;
+package org.hyperledger.besu.ethereum.blockcreation.pluginadapter;
 
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Log;
