+++ platform/build.gradle
@@ -58,6 +58,7 @@ dependencies {
     api project(':ethereum:api')
     api project(':ethereum:api:plugin-api')
     api project(':ethereum:blockcreation')
+    api project(':ethereum:blockcreation:plugin-api')
     api project(':ethereum:core')
     api project(':ethereum:core:plugin-api-chain')
     api project(':ethereum:core:plugin-api-execution')
