+++ plugin-api/build.gradle
@@ -41,6 +41,7 @@ dependencies {
   api project(':ethereum:core:plugin-api-chain')
   api project(':ethereum:core:plugin-api-worldstate-backend')
   api project(':ethereum:core:plugin-api-execution')
+  api project(':ethereum:blockcreation:plugin-api')
   api project(':datatypes')
   api 'org.apache.commons:commons-lang3'
   api 'io.consensys.tuweni:tuweni-bytes'
@@ -80,7 +81,8 @@ tasks.register('checkAPICompatibility', me.champeau.gradle.japicmp.JapicmpTask) 
     project(':ethereum:api:plugin-api').tasks.named('jar'),
     project(':ethereum:core:plugin-api-chain').tasks.named('jar'),
     project(':ethereum:core:plugin-api-worldstate-backend').tasks.named('jar'),
-    project(':ethereum:core:plugin-api-execution').tasks.named('jar'))
+    project(':ethereum:core:plugin-api-execution').tasks.named('jar'),
+    project(':ethereum:blockcreation:plugin-api').tasks.named('jar'))
   newClasspath.from(sourceSets.main.runtimeClasspath)
 
   packageIncludes = [
