+++ settings.gradle
@@ -42,6 +42,7 @@ include 'datatypes'
 include 'ethereum:api'
 include 'ethereum:api:plugin-api'
 include 'ethereum:blockcreation'
+include 'ethereum:blockcreation:plugin-api'
 include 'ethereum:core'
 include 'ethereum:core:plugin-api-chain'
 include 'ethereum:core:plugin-api-execution'
